Renaming file bool.rs to boolean.rs to prevent bugs with primitive type bool
